fixed mouse movement

+++ scenes_inf443/project/src/camera.hpp
@@ -15,7 +15,7 @@ struct camera_controller_custom : camera_controller_first_person_euler
 
 	float mouse_sensitivity = 3.0;
 	float keyboard_sensitivity = 3.0;
-	double const pi = std::atan(1) * 4;
+	float const pi = std::atan(1) * 4;
 
 	camera_controller_custom();
